Rename vpc http files to reflect class names.
