+++ lib/ibm/cloud/sdk/vpc/http_vpc.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 require 'ibm/cloud/sdk_http'
-require_relative 'http_vpc/vpc_http'
+require_relative 'http_vpc/base_http'
 
-require_relative 'http_vpc/vpc_collection'
-require_relative 'http_vpc/vpc_instance'
-
+require_relative 'http_vpc/base_collection'
+require_relative 'http_vpc/base_instance'
