+++ lib/ibm/cloud/sdk/vpc/http_vpc/base_http.rb
